Several casting-related fixes.

+++ include/di/systems/display/opengl_window.hpp
@@ -3,6 +3,7 @@
 
 #include <array>
 #include <cstddef>
+#include <cstdint>
 #include <stdexcept>
 #include <string>
 #include <utility>
@@ -59,27 +60,27 @@ public:
     SDL_GL_SwapWindow(native_);
   }
 
-  bool                       current      ()                           const
+  bool                         current      ()                           const
   {
     return SDL_GL_GetCurrentWindow() == native_ && SDL_GL_GetCurrentContext() == opengl_context_;
   }
-  void                       set_current  ()                           const
+  void                         set_current  ()                           const
   {
     SDL_GL_MakeCurrent(native_, opengl_context_);
   }
-  opengl_swap_mode           swap_mode    ()                           const
+  opengl_swap_mode             swap_mode    ()                           const
   {
     if (!current()) set_current();
     return static_cast<opengl_swap_mode>(SDL_GL_GetSwapInterval());
   }
-  void                       set_swap_mode(opengl_swap_mode swap_mode) const
+  void                         set_swap_mode(opengl_swap_mode swap_mode) const
   {
     if (!current()) set_current();
     SDL_GL_SetSwapInterval(static_cast<int>(swap_mode));
   }
-  std::array<std::size_t, 2> drawable_size()                           const
+  std::array<std::uint32_t, 2> drawable_size()                           const
   {
-    std::array<std::size_t, 2> drawable_size;
+    std::array<std::uint32_t, 2> drawable_size;
     SDL_GL_GetDrawableSize(native_, reinterpret_cast<int*>(&drawable_size[0]), reinterpret_cast<int*>(&drawable_size[1]));
     return drawable_size;
   }
